Moved services related to geolocating visits to the Visit\Geolocation namespace
