+++ module/CLI/test/Command/Visit/LocateVisitsCommandTest.php
@@ -15,9 +15,9 @@ use Shlinkio\Shlink\Core\Entity\Visit;
 use Shlinkio\Shlink\Core\Entity\VisitLocation;
 use Shlinkio\Shlink\Core\Exception\IpCannotBeLocatedException;
 use Shlinkio\Shlink\Core\Model\Visitor;
-use Shlinkio\Shlink\Core\Visit\VisitGeolocationHelperInterface;
-use Shlinkio\Shlink\Core\Visit\VisitLocator;
-use Shlinkio\Shlink\Core\Visit\VisitToLocationHelperInterface;
+use Shlinkio\Shlink\Core\Visit\Geolocation\VisitGeolocationHelperInterface;
+use Shlinkio\Shlink\Core\Visit\Geolocation\VisitLocator;
+use Shlinkio\Shlink\Core\Visit\Geolocation\VisitToLocationHelperInterface;
 use Shlinkio\Shlink\IpGeolocation\Exception\WrongIpException;
 use Shlinkio\Shlink\IpGeolocation\Model\Location;
 use ShlinkioTest\Shlink\CLI\CliTestUtilsTrait;
