+++ module/Core/src/Visit/Geolocation/VisitToLocationHelperInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\Visit;
+namespace Shlinkio\Shlink\Core\Visit\Geolocation;
 
 use Shlinkio\Shlink\Core\Entity\Visit;
 use Shlinkio\Shlink\Core\Exception\IpCannotBeLocatedException;
